JOB/requirements.txt
# HTTP requests
requests==2.31.0

# HTML parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# PostgreSQL database connection
psycopg2-binary==2.9.9

# Optional: if you later want Selenium or Playwright
# selenium==4.11.2
# playwright==1.40.0
